added recall computation on vector of users

+++ io.cpp
@@ -6,7 +6,7 @@ void writeFVec(std::vector<float> fvec, std::string fName) {
   std::ofstream opFile(fName);
 
   if (opFile.is_open()) {
-    for (unsigned int i = 0; i < fvec.size(); i++) {
+    for (size_t i = 0; i < fvec.size(); i++) {
       opFile << i << " " << fvec[i] << std::endl;
     }
     opFile.close();
